moved stuff to globals.js and functions.js
